Rework crypto support
This commit reworks the crypto support in AsyncSSH to separate packet
encryption and decryption into its own module, rather than implementing
some of the logic directly in the connection module. It also flattens
the directory structure of the modules within the crypto package,
eliminating the "pyca" subdirectory that was created when AsyncSSH used
a mix of PyCA and PyCrypto.

In addition, this commit includes some minor cleanup in the agent module
and moves the load_default_keypairs function back into the public_key
module.