Fix an error related to OpenSSH format arcfour encryption of private keys
The previous crypto refactoring inroduced an error in reading and
writing OpenSSH format private keys encrypted with the arcfour cipher.
I didn't catch this before checkin because I'm running a newer version
of OpenSSH which has removed this cipher, so the OpenSSH
interoperability testing for that was skipped. Thankfully, Travis CI
caught this for me. This commit should fix this issue.