Improve SFTP glob matching
This commit adds support for using '**' to do a recursive directory
search in a glob pattern match. It also adds support for matching on
character ranges like '[A-Za-z0-9]', and adds support for getting only
directories back in a search by ending a glob pattern with a trailing
'/'.

In addition to working in the glob() method of SFTPClient, all of these
new matching capabilities are available in mget(), mput(), and mcopy().