Add debugging output which shows received kexinit algorithms
This commit adds debug logging to show the algorithms listed in a
received kexinit message. Previously, the debug logging showed the
algorithms requested in the sent kexinit and the algorithms that were
actually negotiated, but it did not show the algorithms advertised in
the received kexinit. Now it does, when the debug level is 2 or higher.