Add support for host-based authentication
This commit adds client and server side support for host-based SSH
authentication. The server-side support is probably more interesting
here, allowing AsyncSSH to authenticate connections from all users on
a host without each user needing to have their own keys. If AsyncSSH
is given access to a trusted key for a client host, though, this
support can be used to allow AsyncSSH to connect as any user on that
client.

Potential future work could include supporting ssh-keysign on systems
which have it enabled, to avoid AsyncSSH needing to have direct access
to the host key when acting as a client.
In addition to the above, this commit fixes a minor issue with
enforcing the source-address option in OpenSSH user certificates and
with error handling when canonicalizing usernames with SASLPrep. It
also fixes a couple of doc string issues.