Add support for ssh-keysign in host-based authentication
This commit adds support for accessing client host keys via the OpenSSH
ssh-keysign program when doing host-based authentication. If ssh-keysign
is present and enabled on the system, AsyncSSH can use it instead of
signing the authentication request itself, avoiding the need to have
access to the host private key.

This commit also adds support for using pathlib path objects in calls to
read and write public keys and certificates.