Add support for auth_completed callback in SSHServer
This commit adds support for a new auth_completed callback in the
SSHServer class which is called as soon as authentication is completed
successfully on each new connection. This can be used to set up state
based on the username or on options in the authorized_keys list or
certificate associated with the user before any requests are processed
on the connection.