Change channel Unicode handling to use incremental codec
This commit changes the AsyncSSH channel code to use the incremental
encoder and decoder classes in the codecs module rather than doing
its own buffering of partial Unicode data.

This commit also improves the information returned when a Unicode
decode error occurs, and avoids the caller seeing both the original
UnicodeDecodeError and a DisconnectError when this occurs. As part of
the better error handling, this commit also simplifies the way cleanup
is done when a disconnect error is raised.

In addition, this commit delays processing of incoming data on a channel
until after the channel and session objects are returned to the
application opening the channel. Previously, it was possible for some
incoming data to be delivered to the session object even before the
channel open call returned.

Unit tests have also been adjusted to take into account these changes,
and some general cleanup was done there.