Add support for setting Unicode error handling strategy
This commit adds support for an "errors" argument in conjunction with
"encoding" when creating new SSH sessions/streams/processes. This error
handling will be applied on any data exchanged on the SSH channels
associated with these objects.

In addition, the session_encoding argument in create_server() now has a
corresponding session_error argument which will be used as the error
handling strategy for all newly created sesions on that server.

When working with SSH processes, the error handling strategy set on the
process will also automatically be used when doing I/O redirection on
that process.

Finally, the get_comment() and set_comment() functions on private/public
keys and certificates now support an "errors" argument along with
"encoding" when getting/setting comments using Unicode strings.