Add explicit handlers for OpenSSH keepalive global & channel requests
This commit adds explicit handlers for OpenSSH keepalive global and
channel request messages, to avoid reporting an unknown request debug
message. These requests are still logged, but they are no longer labeled
as unknown and they are logged at debug level 2 instead of level 1, so
by default they will not pollute the log output.

This commit also fixes a missing "return" in one of the test functions.
It was harmless before, but the new test needed a proper return value to
be passed through.