Add set_extra_info method on SSH connections and channels
This commit adds a set_extra_info method on the connection and channel
subclasses, allowing applications to store additional information on
these objects which can be retrieved just like the other information
available through get_extra_info. As with existing information, anything
set on the connection is inherited by all channels on that connection,
and channel information is also available via the stream reader & writer
classes.