Add get_extra_info method to SSH process classes
This commit adds a get_extra_info() method on the SSHClientProcess and
SSHServerProcess classes, allowing information associated with the
channel the process is associated with to be retrieved from the process
class directly, rather than having to access it through the "channel"
property.

This commit also updates the examples to show this more direct access.