Avoid overflow on systems with 32-bit time_t value
This commit updates the handling of time values in X.509 certificates to
avoid overflow on systems with a 32-bit time_t. While X.509 certificates
should allow dates up to 9999-12-31, systems with a small time_t have
problems with this.

In addition to artificially limiting time values to avoid a time_t
overflow in calls made directly by asyncssh, this commit also avoids a
similar problem seen in calls made from within the cryptography library
on datetime values passed into it.