Add support for parallel read and write requests in SFTPClientFile
This commit adds support in SFTPClientFile to automatically break up
large read or write requests into multiple smaller requests issued in
parallel to the SFTP server, similar to what was already supported for
the higher-level get, put, and copy methods.

This allows very large I/O operations to be performed directly on files
on the SFTP server without having to first put the content in a local
file, and it gets around limitations on some SFTP servers on the maximum
SFTP message size when writing and on the maximum data they'll return in
a single read response.

This feature can be controlled using new block_size and max_requests
arguments in the SFTPClient open() call. If block_size is set to None,
the prior behavior of issuing only a single read or write request is
used, but by default any read or write larger than 16 KB will be
parallelized, with up to 128 simultaneous outstanding read or write
requests.

As part of this change, calls to read() with no size will now query the
remote file size and trigger this same parallel read behavior if it is
enabled. Previously, such read() calls did break the read into multiple
requests, but the requests were issued serially. This should provide
a significant speed improvement when there's a long latency between the
SFTP client and server.

The new max_requests argument is now also available on the SFTPClient
get, put, copy, mget, mput, and mcopy methods.