Update prior fix to SFTP exception handling
This commit updates the previous fix to SFTP exception handling to avoid
showing a "Connection lost" message in log output  when the SFTP connection
is cleanly closed. The updated code still delivers an exception to all
outstanding requests in this case, but the logs now properly report the
clean close in the "exited" messages.