Add the ability to pass connect() keyword arguments from scp()
This commit allows you to forward keyword arguments passed in to
asyncssh.scp() to any asyncssh.connect() calls it makes, allowing
things like credentials to be specified. This can also be used to
specify an alternate port number without the need to use the more
cumbersome (but more flexible) tuple syntax in the scp() call.