Fix a problem in preserving directory structure during recursive scp
This commit fixes a problem where directory structure was not properly
preserved when acting as a sink for a recursive scp transfer. When
leaving a directory, the path was not properly reset back to the parent
directory, causing some files to be written to the wrong location during
the copy.