Add support for "soft" EOF when line editor is enabled
This commit changes the way Ctrl-D is handled in the line editor to
allow EOF to be signalled multiple times on a channel. Previously,
hitting Ctrl-D in the line editor would make it appear as if the
underlying SSH channel had gotten an EOF, causing the stream-based
API to mark the channel as half-closed. Now, hitting Ctrl-D will
result in a read() or readline() call returning an empty string like
it does for a true channel EOF, but after that additional data can
still be received, like what happens when a break or signal is sent.
This should allow applications to continue to use a channel even
after a client hits Ctrl-D to signal EOF for one part of their input.