Disable host certificate validation when known_hosts is None
This commit disables all validation of server host certificates when
known_hosts is set to None. Previously, validation of host keys and CA
keys used to sign server certificates was disabled, but other checks
against the certificate type, validity period, and principals were still
being checked.

Now, when known_hosts=None is specified, all host key _and_ certificate
checks are disabled, allowing connection to a server even if it has an
invalid certificate. This should be used with care -- any use of
known_hosts=None (before or after this change) leaves open the
possibility of man-in-the-middle attacks!