Switch curve25519 DH key exchange to PyCA
This commit switches the curve25519 key exchange to use the
implementation now available in the PyCA (cryptography) package rather
than libnacl, making it available on systems where libnacl and
libsodium are not installed.

For now, the ed25519 support is still based on libnacl, but I'll look at
switching that over once that support has been added to cryptography.