Add support for SSHSubprocessProtocol/SSHSubprocessTransport
This commit adds new subprocess transport and protocol classes
which provide compatibility with asyncio.SubprocessProtocol and
asyncio.SubprocessTransport. Code which is designed to call
loop.subprocess_shell() or loop.subprocess_exec() should be
easily adapted to start a process on a remote system by calling
SSHClientConnection.create_subprocess() instead.

This commit also adds support for using integer signal values or
constant from the "signal" module in the send_signal() method on
SSHClientChannel and SSHClientProcess, and a get_returncode()
call on SSHClientChannel and returncode protocol on SSHClientProcess
which is designed to be compatible with subprocess.Popen.returncode
and its asyncio equivalent.