Add support for sending keepalive messages
This commit adds support for sending keep-alive messages when the SSH
connection is otherwise idle for some period of time. It also allow you
to set a threshold for how many keep-alive messages can be sent without
getting a response before the connection is closed, providing a way to
detect and clean up a connection which is no longer responsive.

This feature can be enabled using the arguments keepalive_interval and
keepalive_count_max in create_connection() and create_server(), or can
be adjusted during the lifetime of a connection by calling the
set_keepalive() method on it.