Ignore errors when loading unsupported key types from default location
This commit changes the loading of private keys from the default
location to ignore unkown key types, preventing a failure from being
reported if there's an Ed25519 key in the default location but AsyncSSH
is installed without Ed25519 key support. This check should also apply
to other other key types in the future as the default key list is
extended.