Add subclasses of DisconnectError
This commit adds a set of subclasses of the DisconnectError exception
and uses them as a more concise way to raise these exceptions as well as
to be more specific about what exceptions are expected in the unit
tests.

This commit also updates the logging around connection_lost() to remove
"Disconnect Error" from exception message text  and better distinguish
betewen connections which are lost at the transport level or due to a
timeout vs. an error of some kind triggering the connection close.