Optimize SFTP glob matching
This commit reduces number of calls to listdir() if glob pattern
contains directory names without glob characters. This could improve
performance significantly if the network latency is high and/or the
intermediate directories have big number of entries.

For example before this commit the pattern "/usr/share/doc/qt5/global/*"
resulted in six listdir() calls but with this commit only one call is
needed. And the total number of returned entries reduced from more
than 1.5K to just couple of dozen in my case.

An additional bonus is that with this commit the case when mount point
of the auto-mounted directory is used without glob characters in the
pattern then it is processed correctly.