Add channel, connection, and env properties to SFTPServer class
This commit adds new properties of connection, channel, and env to the
SFTPServer class, allowing an SFTPServer instance to access the channel
it is associated with, the connection containing that channel, and the
environment variables set by the client when the SFTP session was
opened.

This commit also adds a new env argument to the start_sftp_cient method
of SSHClientConnection, to allow environment variables to be set on the
SFTP client channel which is being opened.

For backward compatibility with existing code, the constructor of
SFTPServer still takes a "conn" argument. However, it is no longer
necessary to explicitly store this value if you need to use it later,
as it is available as a property named "connection" on the SFTPServer
object.