Fix bug in SFTP parallel I/O file reader
This commit fixes a bug in the SFTP parallel I/O reader where some
combinations of read sizes and block sizes would cause a read at the end
of the file to report EOF prematurely. Thanks go to David G for reporting
this problem and providing a reproducible test case!