Fix test failures seen on Fedora Rawhide
This commit updates some test cases associated with encrypted private
keys which were triggering unknown OpenSSL errors on Fedora Rawhide.
Versions of OpenSSL on other platforms don't seem to trigger this
error, but the change here should avoid running into the issue on
any platforms that do. Thanks go to Georg Sauthof for reporting the
issue and providing a test envrionment to help debug it!