Make a copy of the client keys coming from SSHClientConnectionOptions
This commit makes sure to make a copy of the client keys and client host
keys coming from the SSHClientConnectionOptions when creating a new
SSHClientConnection. That way, the original list of keys will remain
available even after the options are used to authenticate on a connection.
Without this fix, attempts to reuse an SSHClientConnectionOptions object
more than once can run into authentication failures.