Improve path handling when copying files on Windows
This commit makes the AsyncSSH SFTP client use the local os "basename"
call when parsing path arguments passed into the various get/put/copy
functions, better handling native Windows paths with drive letters and
backslashes. Previously, posixpath was used for this, which meant that
forward slash needed to be used as a path separator even on Windows for
it to properly construct remote paths from local paths in a copy.