Handle SFTP file reads and copies where a read returns partial data
This commit enhances the SFTP parallel I/O for file reads and file
copies to handle the case where a read returns less data than what was
requested when not yet at end of file. This can happen if the server has
a limit on its read buffer size (such as OpenSSH's limit of 64 KB).

With this change, AsyncSSH will automatically adapt when a read returns
less data than requested, issuing additional reads as necessary to get
all of the data requested. As before, multiple reads are started in
parallel at the requested block_size, but each of those will be turned
into multiple consecutive reads if less than the full block size is
returned.

This change will also now raise an SFTPError if a read issued during a
file copy returns EOF. This should normally not happen, since the file
size is determined when starting the copy and all reads are kept within
that range. However, if the source file is truncated during the copy and
reads beyond that point haven't completed, an error is raised to report
that the file copy was incomplete.

Note that modifications to a source file during a copy will not always
be detected by this new check. If the file size remains unchanged, no
error will be raised even though the copied file could be a mix of old
and new data.  Also, even when the source file changes size, an error may
not be raised if reads for data later in that file completed before the
file size was reduced.