Simplify agent code and avoid warning when default agent is unavailable
This commit cleans up a few things in the SSH agent client code, and
avoids printing a warning message on Windows when an attempt to connect
to the agent using the default path fails. A warning will still be printed
if agent_path or SSH_AUTH_SOCK is explicitly set and the connection fails,
but not when these values are defaulted.