Fix gssapi-with-mic auth to work with GSS key exchange
This commit fixes gssapi-with-mic authentication to work correctly even
when a GSS key exchange methods was used. Normally, this combination
doesn't occur, as gssapi-keyex authentication is usually used instead
of gssapi-with-mic in that case. However, if gssapi-keyex is disabled
on the server, this commit fixes a problem where the GSS context wasn't
reset properly before performing auth.