Make connect_ssh and connect_reverse_ssh methods async context managers
This commit marks the connect_ssh() and connect_reverse_ssh() methods
in SSHClientConnection as async context managers, so their result can
be used directly in "async with" statements. Previously, the main
connect() method was marked this way and could support tunnelled SSH
connections with an optional "tunnel" argument, but "async with" didn't
work properly when using the connect_ssh() call which provides a nicer
looking syntax.