Fix issues with peer address and host name on tunneled SSH connections
This commit fixes an exception in known hosts matching caused by the
peer address potentially being set to a hostname for tunneled SSH
connections. Since no address lookup is done on the client in such
cases, the address cannot be used for known hosts matching, and so
the code now leaves the address unset, unless the destination of the
tunnel is a numeric address.

This commit also better handles setting the target host name to the
remote system when a reverse SSH listener is set up over an SSH tunnel.
The reported hostname in this case will be whatever the tunnel reports
as the original host on the inbound forwarded TCP connection, and this
will be the host and port used to perform the known_hosts lookup.