Make SSHAgentClient capable of being an async context manager
This commit makes SSHAgentClient capable of acting as an async context
manager, automatically closing the connection to the agent when leaving
an "async with" block.

When cleaning up the agent connection, wait_closed() is now called on
Python releases that support it (3.7 and later), avoiding a false
error report by Python 3.8 about garbage collecting an open stream.