Eliminate use of log level warning in AsyncSSH
This commit downgrades logging calls of level warning in AsyncSSH to
debug1, so that nothing is output to the terminal by default. It
really should be up to the application to decide whether these warnings
should be output or not. All the previous messages at this level will
still show up if logging is configured with a level of "debug".