Change SFTPServer constructor to pass channel instead of connection
This commit changes the first argument of SFTPServer constructor from an
SSHServerConnection to an SSHServerChannel. An earlier change in
AsyncSSH 1.16.1 began passing the channel through to SFTPServer to
allow applications to query environment variables set there. However,
to maintain backward compatibility, the constructor was left unchanged
and the channel was passed in after the SFTPServer was constructed.

Now that AsyncSSH is moving to a new major version, the constuctor has
been updated to directly pass in the channel. Applications which subclass
the SFTPServer class and implement an __init__ method will need to be
updated to account for this change and pass the new argument through
to the SFTPServer parent class. If the subclass has no __init__ and
uses only the connection, channel, and env properties of SFTPServer
to access this information, no changes should be needed.