Restore ability to use SSHListener as an async context manager
This commit restores the ability for SSHListener objects to be used as
async context managers. This was an undocumented and unexpected side effect
of previously having that class inherit from asyncio.AbstractServer
and was accidentally removed in 2.0.0. It seems like a useful thing to
continue to support, so it has been added back here.

Using these objects as context managers will cause the listener to be
shut down when the context is exited, but any already opened connections
will remain open.