Fix edge cases in certificate loading
This commit fixes an issue where AsyncSSH didn't properly handle reading
of OpenSSH format certificates when a newline at the end of the line
was missing.

In addition, multiple certificates can now be loaded from a single byte
string value passed to load_certificates(), making it more consistent
with the way multiple certificates can be loaded from a file passed
in as a filename.