Add support for registered key handlers to trigger signals
This commit extends the register_key() support when the line editor is
enabled to allow handlers to trigger delivery of a signal on stdin by
returning a tuple of a signal name and a negative cursor position. The
current input line will remain unchanged, but the requested signal will
be delivered before any other input is processed, allowing "hot keys" to
be defined that trigger actions on the server unrelated to editing the
input line.