Wait for agent to close to avoid Python 3.8 garbage collection warning
This commit tries to avoid a false Python 3.8 warning about not properly
closing the stream associated with an SSH agent connection when a
connection is closed. This stream was closed, but the error can appear
if the stream object is garbage collected before the connection close
has a chance to finish.