Extend SSHProcess redirection to support asyncio streams
This commit extends the SSHProcess redirect mechanism to accept asyncio
StreamReader and StreamWriter objects, allowing asyncio streams to be
plugged in as stdin/stdout/stderr in an SSH process. Among other things,
this allows stdin, stdout, and stderr of asyncio.subprocess.Process
objects to be connected directly to SSH processes.