Check in first cut at support for U2F/FIDO2 security keys
This commit adds support for OpenSSH-compatible U2F and FIDO2 security
keys in AsyncSSH. It has been tested against openssh-portable HEAD,
but since that is not part of an official release, this implementation
may need changes to remain compatible with whatever official OpenSSH
release this feature ships in. In the meantime, please enjoy an early
preview of this functionality.

Features in this commit include:

  * ECDSA and Ed25519 signature algorithms
  * Generation of new keys, with an option to require touch or not
  * Generation of certificates, both as a key being signed and a CA key
  * Read and write of OpenSSH-compatible key and certificate file formats
  * Access to and management of keys loaded in an OpenSSH ssh-agent
  * Compatibility with OpenSSH "middleware" provider API (libsk-libfido2)

In addition to supporting these key types for user keys and certificates,
this implementation also supports them for host keys and certificates.
However, since OpenSSH doesn't have this support yet, AsyncSSH must be
both the client and server to take advantage of this (at least for now).