Allow specifying a certificate to use with an existing keypair
This commit allows you to update the SSH certificate used with an
existing keypair. This can be used if you want to associate a
certificate with a keypair tied to an ssh-agent key which wasn't
imported with the certificate you want to use. It can also be used
to associate a certificate with a key on a smart card, which
doesn't currently allow a certificate to be imported in ssh-agent.