Allow load_keypairs() to read public key files
This commit changes the load_keypairs() function to support reading
public key files in addition to certificates. This can be triggered
by passing in a tuple of (priv_key, pub_key), or the public key can
be found automatically when the private key is read from a file,
by appending ".pub" to the private key filename.

Right now, the only data extracted from the public key file is the
comment, and it will be set as the comment on the resulting keypair
object if there's no comment set in either the private key file or
certificate. This can be useful when the private key is written out in
PKCS1 or PKCS8 format which don't support comments, but the public key
is written out in OpenSSH or RFC4716 format.

Checking is done to make sure the public and private key files refer
to the same key before it is used. If not present or not readable,
the public key file is ignored.