Fix an issue in the SCP server related to handling files with spaces
This commit fixes a problem handling filenames with spaces in the SCP
server code. Thanks go out to Stephen Copeland for reporting the issue
and suggesting a fix!

This also includes a fix for an issue related to handling of filenames
with colons in them passed as byte strings. This issue was previously
fixed when passing in regular strings, but the byte string case was
missed.