Update U2F/FIDO2 security key support to use Python fido2 module
This commit updates the AsyncSSH U2F/FIDO2 security key support to be
based on the Python fido2 module, rather than the "libsk-libfido2"
middleware library which was originally used by OpenSSH. That library
was folded into the OpenSSH source code and is no longer available to
install separately.

One benefit to this is that the necessary FIDO2 library support can now
be installed with 'pip', and AsyncSSH now has a 'fido2' optional extra
which will install FIDO2 support when AsyncSSH is installed.

This new support should continue to provide support for both U2F (CTAP
version 1) and FIDO2 (CTAP version 2) keys, and for both the ECDSA and
Ed25519 signature algorithms. Both user and host keys are supported,
as are OpenSSH certificates based on these keys. As before, support is
also available for using and managing keys via an OpenSSH 8.2 or later
ssh-agent.

Interoperability testing has been done against OpenSSH 8.2, where this
support has now been released. However, AsyncSSH doesn't yet support the
"resident key" feature found in OpenSSH 8.2 -- this should be coming in
a future commit.