Fix an issue with resuming reading when readuntil() returns incomplete read
This commit fixes a problem where reading is not resumed when
readuntil() fails to find a separator match and ends up raising
IncompleteReadError with a partial result. With this change,
readuntil() can be called again after the exception to continue
reading data, until EOF or a signal is received.

Note: Multi-character separators could be missed if they happen to be
split across the partial results which are returned.