Protect client against connecting to non-compliant SSH servers
This commit adds support for a login_timeout client option, to limit
the amount of time the AsyncSSH client will wait for the initial SSH
handshake and authentication to complete. It defaults to two minutes,
just like the existing server-side login_timeout option.

In addition, this commit adds limits on how long banner lines sent
prior to the server version string can be, and on the number of such
 banner lines. This will allow the client to abort more quickly if it
gets a large amount of output that doesn't match what it is expecting
to get from an SSH server.