Change SSHConnection to inherit from asyncio.Protocol for type checkers
This commit makes SSHConnection a subclass of asyncio.Protocol, to make
type checkers happy.

Note: The signature on data_received() in SSHConnection is slightly
different than the one in asyncio.Protocol. It takes an extra optional
argument of datatype, to support tunneled SSH connections. Due to this
change, it was necessary to disable the arguments-differ warning from
pylint.