Add timeout to SSHClientProcess.wait and SSHClientConnection.run
This commit adds an optional timeout argument in the
SSHClientProcess.wait() and SSHClientConnection.run() methods, allowing
the caller to specify the maximum time for the process to exit. If this
time is exceeded, a TimeoutError exception is raised which contains
the process information, including any output on stdout and stderr
received before the timeout occurred.