Fix exception handling in SFTP parallel I/O class
This commit fixes a problem in the exception handling in the SFTP
Parallel I/O class where a low-level connection failure was not
being handled correctly, looking for a "code" member of the exception
object which is only present for higher-level SFTP errors. Thanks go
to Mikhail Terekhov for reporting this!