Fix SCP to use new SFTPError subclasses
This commit enhances SCP to use the new subclasses of SFTPError
introduced in a prior commit. Previously, SCP errors raised in
the SFTP classes would have returned these subclasses, but some
of the errors raised within the SCP implementation itself were
still returning the parent SFTPError class. This has now been
corrected.

As part of this change, the internal "SCPError" subclass of SFTPError
has been eliminated completely. This class was not part of the public
AsyncSSH API and applications should not have been using it, so this
is not a compatibility break. Application code should only have been
using SFTPError (or its new subclasses). This should reduce confusion
in debugging output, where the internal SCPError class was sometimes
visible. Now, exceptions will always show up as one of the public
SFTPError subclasses.