Add support for sending and receiving EXT_INFO messages from the client
This commit adds support for accepting and parsing EXT_INFO messages from
SSH clients when AsyncSSH is acting as a server, and for sending an
EXT_INFO message containing the "global-requests-ok" extension when
AsyncSSH is acting as a client and the server supports client extensions.

Since AsyncSSH doesn't currently disable sending global requests based on
the advertised client software and version, it doesn't currently do
anything different when receiving the 'global-requests-ok' extension.
However, this change provides the support to easily parse client extensions
that might be needed in the future.