Fix keepalive handler to avoid leaking a timer in some cases
This commit fixes keepalive support to avoid leaking a timer object if
set_keepalive() is called on a connection object before user
authentication has completed. With this change, the new keepalive
parameters are recorded but the timer itself isn't started until after
authentication, just as it would be when passing n keepalive parameters
when the SSH connection is created. Thanks go to Tom van Neerijnen
for reporting this issue!