Add support for default algorithms and OpenSSH algorithm syntax
This commit changes AsyncSSH to no longer enable all supported algorithms
by default. When algorithms is registered, a flag is passed to indicate
whether the algorithm should be enabled by default or not, allowing
weaker algorithms to be supported but only when an application
explicitly requests them. The API documentation has been updated
to indicate which algorithms are enabled by default.

In addition, this change allows support for OpenSSH-style wildcard
syntax when requesting algorithms. Both positive and negative patterns
can be supported, as an optional prefix of '^', '+', or '-' to add
or remove algorithms from the default list rather than completely
replacing it. This new syntax is supported when algorithms are passed
in as a string value rather than a list of strings.

This new support applies to server_host_key_algs, kex_algs,
encryption_algs, mac_algs, compression_algs, and signature_algs.