Add new byte count and time interval parsers
This commit adds new byte count and time interval parsers that take a
series of numeric values and suffixes, allowing the specification
of kilobytes, megabytes, and gigabytes and time intervals in seconds,
minutes, hours, days, or weeks. These will be used in a follow-up
commit to parse AsyncSSH re-keying byte and time limits when they
are specified as strings.

This commit also changes the X.509 certificate generation code to use
this new time interval parser for parsing relative times used for
setting certificate validity periods.